moves files to the corresponding folders
