+++ lib/presentation/home/screens/screens.dart
@@ -1,0 +1,4 @@
+// Barrel file
+export 'package:banking_flutter_app/presentation/home/screens/buttons/buttons_screen.dart';
+export 'package:banking_flutter_app/presentation/home/screens/cards/cards_screen.dart';
+export 'package:banking_flutter_app/presentation/home/screens/home/home_screen.dart';
