--- lib/presentation/screens/screens.dart
@@ -1,4 +1,0 @@
-// Barrel file
-export 'package:banking_flutter_app/presentation/screens/buttons/buttons_screen.dart';
-export 'package:banking_flutter_app/presentation/screens/cards/cards_screen.dart';
-export 'package:banking_flutter_app/presentation/screens/home/home_screen.dart';
